init db folder for database
