+++ .gitignore
@@ -11,3 +11,4 @@ package-lock.json
 # System Files
 .DS_Store
 Thumbs.db
+/.env
