+++ package.json
@@ -23,6 +23,8 @@
   "dependencies": {
     "axios": "^0.19.2",
     "cors": "^2.8.5",
+    "body-parser": "^1.19.0",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "pg": "^8.2.1",
     "xml-js": "^1.6.11"
@@ -33,7 +35,6 @@
     "@types/express": "^4.17.6",
     "@types/node": "^14.0.5",
     "@types/pg": "^7.14.3",
-    "body-parser": "^1.19.0",
     "nodemon": "^2.0.4",
     "typescript": "^3.9.3"
   }
